Rename Document namespace from Common to MarketDocument (#133)
