+++ source/GreenEnergyHub.TimeSeries/source/GreenEnergyHub.TimeSeries.Domain/GreenEnergyHub.TimeSeries.Domain.csproj
@@ -30,7 +30,7 @@ limitations under the License.
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="Common" />
+      <Folder Include="MarketDocument" />
       <Folder Include="Messages" />
     </ItemGroup>
