+++ source/GreenEnergyHub.TimeSeries/source/GreenEnergyHub.TimeSeries.Domain/MarketDocument/DocumentType.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace GreenEnergyHub.TimeSeries.Domain.Common
+namespace GreenEnergyHub.TimeSeries.Domain.MarketDocument
 {
     /// <summary>
     /// The document type indicates the intended business context of this business message.
